src/main/java/net/sf/brightside/travelsystem/metamodel/beans/MealBean.java
package net.sf.brightside.travelsystem.metamodel.beans;

import java.io.Serializable;
import java.util.List;

import javax.persistence.Entity;
import javax.persistence.OneToMany;

import net.sf.brightside.travelsystem.core.beans.BaseBean;
import net.sf.brightside.travelsystem.core.beans.Identifiable;
import net.sf.brightside.travelsystem.metamodel.Meal;
import net.sf.brightside.travelsystem.metamodel.Reservation;

@Entity
public class MealBean extends BaseBean implements Meal, Serializable,
		Identifiable {

	private static final long serialVersionUID = 1L;

	private List<Reservation> reservations;
	private String name;

	public String getName() {
		return name;
	}

	public void setName(String name) {
		this.name = name;
	}

	@OneToMany(mappedBy = "meal", targetEntity = ReservationBean.class)
	public List<Reservation> getReservations() {
		return reservations;
	}

	public void setReservations(List<Reservation> reservations) {
		this.reservations = reservations;
	}

	@Override
	public String toString() {
		return this.getName();
	}

}
